+++ src/Main/Components/Calibration/UiCalibrationCheck.cpp
@@ -1,5 +1,5 @@
 #include "UiCalibrationCheck.h"
-#include "MainWindow.h"
+#include "src/Main/MainWindow.h"
 
 #include "src/Basics/QtBasics.h"
