+++ src/Main/Components/DefaultControls/UiMotionControl.cpp
@@ -1,5 +1,5 @@
 #include "UiMotionControl.h"
-#include "MainWindow.h"
+#include "src/Main/MainWindow.h"
 
 #include <QDebug>
