+++ src/Main/MainWindow.h
@@ -5,10 +5,10 @@
 #include <QMainWindow>
 #include <QTranslator>
 
-#include "UiCustomActions.h"
-#include "UiSerialHandler.h"
-#include "UiCalibrationCheck.h"
-#include "UiMotionControl.h"
+#include "Components/Calibration/UiCalibrationCheck.h"
+#include "Components/Serials/UiSerialHandler.h"
+#include "Components/DefaultControls/UiMotionControl.h"
+#include "Components/CustomCmds/UiCustomActions.h"
 
 
 namespace Ui {
